Product refactor

+++ src/product/variation/variation.service.ts
@@ -1,7 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { DeleteFileDto } from 'src/dto/deleteFilde.dto';
 import { ProductVariationDto } from 'src/dto/productVariation.dto';
 import { ProduktVariations } from 'src/entity/produktVariations';
+import { PhotoService } from 'src/service/photoService';
 import { Repository } from 'typeorm';
 
 @Injectable()
@@ -9,6 +11,7 @@ export class VariationService {
     constructor(
         @InjectRepository(ProduktVariations)
         private produktVariationsRepository: Repository<ProduktVariations>,
+        private photoService: PhotoService
     ) {}
 
     async findAllforSelect() {
@@ -57,7 +60,15 @@ export class VariationService {
 
     async delete(sku: string) {
         try {
-            return await this.produktVariationsRepository.delete(sku);
+          const item = await this.produktVariationsRepository.findOne({ where: { sku: sku}});
+
+          if(item.quanity > 0)
+            throw new HttpException('Item kann nicht gelöscht werden, Menge is größer als 0', HttpStatus.BAD_REQUEST)
+
+          if(item.image && item.image.length > 2)
+           await this.deleteImage({produktid: sku, fileid: item.image});
+
+           return await this.produktVariationsRepository.delete(sku);
         } catch (err) {
             return err;
         }
@@ -71,5 +82,42 @@ export class VariationService {
             return err;
         }
        
+    }
+    async deleteImage(image: DeleteFileDto) {
+      try {
+        const item = await this.produktVariationsRepository.findOne({where: { sku: image.produktid }});
+        
+        if(!item)
+          return false;
+  
+        //  const images: string[] = JSON.parse(item.image);
+        // const index = images.findIndex((tmp) => tmp === image.fileid)
+        //  images.splice(index, 1);
+        //  item.image = JSON.stringify(images);
+        this.photoService.deletePhoto(image);
+        item.image = '';
+        await this.produktVariationsRepository.save(item);
+        
+        return true;
+      } catch (err) {
+        return err;
+      }
+    }
+    async addImage(image: string, productid: string): Promise<boolean> {
+      try {
+        const item = await this.produktVariationsRepository.findOne({where: { sku : productid }});
+        if(!item)
+          return false;
+  
+      //  const currentImages: string[] = JSON.parse(item.image);
+      //  currentImages.push(image);
+      //  item.image = JSON.stringify(currentImages);
+        item.image = image;
+        await this.produktVariationsRepository.save(item);
+  
+        return true;
+      } catch (err) {
+        return err;
+      }
     }
 }
